perf: compact CacheEntry state, halve HeavyKeeper counters, and add O(1) RollingWindow sum (master)
CacheEntry now stores the logical metadata compactly: decisionEpoch (56 bits) and keyState (2-bit code) are packed into a single packedState long, while hardTtlMs, softTtlMs, normalHardTtlMs and normalSoftTtlMs are encoded as ints with a 2-bit unit selector and a 30-bit mantissa. TTLs up to 2^30 − 1 ms round-trip exactly; longer values fall into second, minute, or hour tiers with sub-unit error; Long.MAX_VALUE remains a dedicated infinite sentinel. The constructor enforces the ADR-0019 invariant that isVersionDegraded equals dataVersion < 0, rejects decisionEpoch values beyond 56 bits, and rejects negative or out-of-range TTLs. withIsVersionDegraded is gone; withDataVersion now derives the flag from the new version's sign, and toBuilder is handwritten because the logical fields no longer map one-to-one to Lombok-generated fields.

HeavyKeeper's per-slot windows and slotSums were changed from long[] to int[], and member Node counts from AtomicLong to AtomicInteger, halving the sketch footprint and improving cache locality. External increments are clamped to Integer.MAX_VALUE, applyIncrement saturates instead of wrapping, decays still use long intermediates to avoid overflow in proportional subtraction, and persisted snapshot counts are clamped when warmed. RollingWindow now keeps a padded O(1) running total: add mirrors each bucket write and tick subtracts the evicted value via getAndSet, so sum becomes a single volatile read instead of an O(windowSize) scan without changing the approximate window semantics.

ZetaSchedulingConfiguration no longer depends on BroadcastBuffer or CacheSyncPublisher; it now accepts an Optional<Cache<String,Object>> L1 cache and schedules Caffeine cleanUp every five seconds, so expired entries are physically reclaimed even when the cache is idle. The DefaultEvaluator CV history was reduced from 32 to 16 windows, halving per-key buffer memory while keeping enough samples for stable decisions.

ExpireManagerImpl refreshes now use withDataVersion only, relying on the sign-derived degraded flag. The remaining changes are test and doc updates: CONFIG.md and CONFIG.zh.md document the new L1 cleanup interval, and the sync/version tests now model degraded versions with negative numbers per the new invariant.

+++ common/src/test/java/io/github/hyshmily/zeta/hotkeydetector/heavykeeper/CurMeasurementTest.java
@@ -31,8 +31,8 @@ class CurMeasurementTest {
       for (long i = 0; i < totalOps; i++) {
         hk.addDirect(key, 1);
 
-        long[] slotSums = (long[]) slotSumsField.get(hk);
-        for (long cur : slotSums) {
+        int[] slotSums = (int[]) slotSumsField.get(hk);
+        for (int cur : slotSums) {
           if (cur > maxCur) maxCur = cur;
         }
